requirements.txt
faiss_cpu==1.7.3
huffman==0.1.2
matplotlib==3.6.2
numpy==1.24.0
opencv_python==4.6.0.66
Pillow==9.3.0
Pillow==9.5.0
scikit_image==0.19.3
scikit_learn==1.2.0
scipy==1.10.1
varname==0.11.1
